Implement protocol on the client side.
Supports cancellable automatic reconnection, and should be abstracted
enough for us to be able to swap out transport/encoding if needed.

Input code stub is unrelated, but useful for early testing, so I'm
leaving it in.

Doesn't do any validation (yet). Not sure if client really needs that.

Heir is a helper library for prototype inheritance. Game.Socket and
Game.WSTransport use events to communicate with the client code,
EventEmitter is boilerplate for that.

Closes #3. Closes #8.